Sync objets with DOM elements and add toggle button
Before, the index was designed via an id which was assigned when each
table was created, this worked fine but if the user decided to delete two
tables consecutively then the id would not correspond with myLibrary.
To fix this I ended up using IndexOf().

Aditionally, before each row had the class of its respective property,
changed it so now only the row of read has an class.